removed config file
