+++ tests/unit/utils/oneview.py
@@ -1,0 +1,40 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+###
+# Copyright (2016-2017) Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+###
+
+from __future__ import (absolute_import, division, print_function)
+__metaclass__ = type
+
+import json
+
+class OneViewClientTest(object):
+
+    @classmethod
+    def from_json_file(cls, file_name):
+        """
+        Construct OneViewClient using a json file.
+
+        Args:
+            file_name: json full path.
+
+        Returns:
+            OneViewClient:
+        """
+        with open(file_name) as json_data:
+            config = json.load(json_data)
+
+        return cls(config)
